Add partially functional website and file parsing
The script can now parse websites and files. For website parsing, the
contents of the site are fetched and parsed for anything matching
possible names. Files are parsed similarly, except they are read line by
line and series of multiple words are inspected to see if they are
possibly a name.

Note that a lot of false positives are possible in this approach.