Rewrite project from Python -> Go
The previous Python implementation was disappointingly slow, especially
with installing on a new machine. Install time took close to 3 minutes,
which is long enough to warrant a new approach. Since the tool is fairly
trivial, it made sense to try rewriting it in a new language from
scratch. The results are much more satisfactory.

Overall the tool is about 5x faster to run, and everything can be
installed in under 1 minute. Note that this does not include any
optimizations at all, purely rewritten logic from Python to Go.

I plan to package the tool for easier install on at least Arch and
Ubuntu in the near future, for the small number of people who might be
interested in using it (likely HTB users).